dissolve NG game into components
